requirements.txt
chromadb
sentence-transformers
anthropic
pypdf2
beautifulsoup4
python-dotenv
